server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.7
flatbuffers==25.12.19
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.46.0
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
onnxruntime-gpu==1.24.1
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.9.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.15
python-dotenv==1.2.1
python-multipart==0.0.22
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.0
starlette==0.52.1
sympy==1.14.0
tifffile==2026.1.28
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
